requirements.txt
blinker==1.6.2
click==8.1.7
Flask==2.3.3
Flask-Cors==4.0.0
flask-swagger-ui==4.11.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
protobuf==4.21.12
Werkzeug==2.3.7
